Relocated new bounty and proposal button to their own components
