+++ components/votes/VotesPage.tsx
@@ -1,13 +1,13 @@
 import { Box, Grid, Typography } from '@mui/material';
 import { VoteStatus } from '@prisma/client';
 import charmClient from 'charmClient';
-import CreatePageButton from 'components/common/Page/CreatePageButton';
 import { CenteredPageContent } from 'components/common/PageLayout/components/PageContent';
 import { filterVotes, sortVotes, ViewOptions, VoteFilter, VoteSort } from 'components/[pageId]/DocumentPage/components/VotesSidebar';
 import { useCurrentSpace } from 'hooks/useCurrentSpace';
 import { usePages } from 'hooks/usePages';
 import { Reducer, useReducer } from 'react';
 import useSWR from 'swr';
+import NewProposalButton from './components/NewProposalButton';
 import VotesTable, { VoteRow } from './components/VotesTable';
 
 export interface ViewState {
@@ -66,7 +66,7 @@ export default function VotesPage () {
         <Grid item xs={12} lg={8} display='flex'>
           <Box gap={3} sx={{ display: 'flex', alignItems: { xs: 'flex-start', lg: 'center' }, width: '100%', justifyContent: { xs: 'flex-start', lg: 'flex-end' }, flexDirection: { xs: 'column-reverse', lg: 'row' } }}>
             <ViewOptions voteSort={viewState.sortBy} voteFilter={viewState.filterBy} setVoteSort={setVoteSort} setVoteFilter={setVoteFilter} />
-            <CreatePageButton type='proposal' />
+            <NewProposalButton />
           </Box>
         </Grid>
       </Grid>
